[Fix] #5 - orderNum 타입 Double로 변경

+++ Memento-iOS/Memento-iOS/Source/Network/ToDoList/DTO/Response/ToDoListResponseData.swift
@@ -21,5 +21,5 @@ struct ToDoListTotalResponseDataTest: Codable, Equatable {
     let tagName: String
     let tagColor: String
     let toDoType: String
-    let orderNum: Int
+    let orderNum: Double
 }
